[FEATURE #10] Add more metrics
As more insight is needed, add more metrics and add support for the new
metrics added in Go 1.16.

 * Renamed internal/conf/statsd*.go -> internal/conf/metrics*.go as more
 types of reporters will probably have to be added in the future.
 * Added code to get the new go metrics. This is mostly the same implementation
 as the docs, but with statsd client instead.

The metrics with the Histogram type is probably not directly suited to this
direct use, but I have attempted to extract value from it, however I am not
sure if this is correct until I can see the metrics reported.

I also elected to keep the old metrics for now, but maybe phase these out
once the new ones have been validated.