[FEATURE] #71 Upgrade vulnerable jwt lib
When I started the task, I discovered that the lib was already updated,
however there was still a couple of cleanups in the authentication setup
that needed to be done, so I did those as well.

The most important change is the cleanup of the error handling in the
validation func, as a failure to validate the PEM kinda swallowed it's
error. It would still fail further down, as the certificate would then be
invalid, however, it is much cleaner to handle it properly.

Also moved the alg check inside the jwt parser instead of doing it myself.