[GIT-65] Add support for different login providers
So, there is a need to support different login providers when communicating
with datalayers through jobs. This feature adds support to add different
providers, and allows using them in a job by referencing the provider name.

Currently only "bearer" (jwt-token) and "basic" types are supported, however
it is backed by an interface, so in theory any type of provider could be supported.

To prevent having to store secrets and passwords in plain text, a ValueReader
type has been added, that provides the ability to read from the environment
and also from AWS SSM.

In theory this could be extended to support also the other default secret
providers that are supported by the Viper package, however, this has not
been attempted, and is left for another day.

Please see the oas3 file for endpoint documentation.

Creating a new basic provider:

POST /provider/logins:
```json
{
    "name":"login1",
    "type":"basic",
    "user: {
        "type": "text",
        "value": "server1"
    },
    "password": {
        "type": "env",
        "value": "SERVER1_SECRET_PASSWORD"
    }
}

```