add id-query flag "details"
when flag is set, datahub will inject "datahub_details" as extra
property into the resulting entity. this is useful for debugging and
understanding how data evolved inside the datahub. "datahub_details"
contains information about accessed datasets and change history per dataset.

the implementation is done in a separate module tree, service, and is
likely to change when more functionality is moved to service modules.
The idea is to replicate functionality in a more decoupled way, while
still keeping the same performance levels.