remove aws-ssm integration
The motivation for this feature removal is that Datahub's ability to pull config parameters from SSM on startup comes with the cost
of 10MB binary size for the aws libraries, which is 25% of total binary size.

Users of this feature are advised to solve the feeding of config from ssm to Datahub utilizing suitable deploy infrastructure
that provides environment variables for Datahub to pick up.