re-implement entity equality check
the new implementation is now correct for nested entities, number
values, array values. see entity_test.go for covered cases.

the new implementation is twice as fast when comparing entities with
only literal values, compared to the old impl. When arrays or nested
entities are encountered, it is comparably slower. but the old impl
would give a wrong result while being fast.

There is a benchmark test included in entity_test.go, which illustrates
the performance difference.